feat(MCP): Add Docker image and release publishing
Ship the MCP server as a Wolfi-based image, built in two stages: uv resolves
dependencies and installs the project as a non-editable wheel into a
relocatable venv, which is copied into a minimal runtime that ships neither
pip nor uv and runs as nobody. The image defaults to HTTP transport bound to
0.0.0.0:8000.

Add a /health endpoint and point the image HEALTHCHECK at it. Share the HTTP
client test fixtures via conftest so both the OAuth and health tests can use
them.

Add a workflow that builds the image (ephemeral on PRs), pushes to GHCR on
main, and publishes to Docker Hub and Quay on release, reusing the existing
reusable docker build/publish workflows.

beep boop
