+++ .github/workflows/mcp-docker-build-publish.yml
@@ -1,0 +1,56 @@
+name: MCP Build Docker Image and Publish
+
+on:
+  pull_request:
+    paths:
+      - mcp/**
+      - .github/workflows/mcp-docker-build-publish.yml
+      - .github/workflows/.reusable-docker-build.yml
+      - .github/workflows/.reusable-docker-publish.yml
+    types: [opened, synchronize, reopened, ready_for_review]
+  push:
+    branches:
+      - main
+    paths:
+      - mcp/**
+      - .github/workflows/mcp-docker-build-publish.yml
+      - .github/workflows/.reusable-docker-build.yml
+      - .github/workflows/.reusable-docker-publish.yml
+  release:
+    types:
+      - released
+
+jobs:
+  docker-build-mcp:
+    name: Build MCP Image
+    uses: ./.github/workflows/.reusable-docker-build.yml
+    with:
+      file: mcp/Dockerfile
+      image-name: flagsmith-mcp
+      # On pull requests, validate the build without pushing to the registry.
+      ephemeral: ${{ github.event_name == 'pull_request' }}
+
+  # Publish to Docker Hub
+
+  docker-publish-mcp:
+    needs: [docker-build-mcp]
+    uses: ./.github/workflows/.reusable-docker-publish.yml
+    if: github.event_name == 'release'
+    with:
+      source-images: ${{ needs.docker-build-mcp.outputs.image }}
+      target-images: flagsmith/flagsmith-mcp
+    secrets: inherit
+
+  # Publish to Quay.io
+
+  docker-publish-quay-mcp:
+    needs: [docker-build-mcp]
+    uses: ./.github/workflows/.reusable-docker-publish.yml
+    if: github.event_name == 'release'
+    with:
+      target-registry-url: quay.io
+      docker-username: ${{ vars.QUAY_PUBLISH_USERNAME }}
+      docker-password-secret-name: QUAY_PUBLISH_PASSWORD
+      source-images: ${{ needs.docker-build-mcp.outputs.image }}
+      target-images: quay.io/${{ vars.QUAY_ORGANISATION_NAME }}/flagsmith-mcp
+    secrets: inherit
