+++ mcp/tests/integration/conftest.py
@@ -1,5 +1,7 @@
 from collections.abc import AsyncIterator
+from typing import Callable
 
+import httpx
 import openapi_pydantic as openapi
 import pytest
 from fastmcp import Client, FastMCP
@@ -8,6 +10,8 @@ from respx import MockRouter
 
 from flagsmith_mcp import config, constants
 from flagsmith_mcp import server as server_module
+
+HTTPClientFactoryFixture = Callable[[FastMCP], AsyncIterator[httpx.AsyncClient]]
 
 
 @pytest.fixture
@@ -55,3 +59,24 @@ def server() -> FastMCP:
 async def client(server: FastMCP) -> AsyncIterator[Client[FastMCPTransport]]:
     async with Client(transport=server) as connected:
         yield connected
+
+
+@pytest.fixture
+def http_client_factory() -> HTTPClientFactoryFixture:
+    async def factory(server: FastMCP) -> AsyncIterator[httpx.AsyncClient]:
+        transport = httpx.ASGITransport(app=server.http_app())
+        async with httpx.AsyncClient(
+            transport=transport, base_url="http://testserver"
+        ) as connected:
+            yield connected
+
+    return factory
+
+
+@pytest.fixture
+async def http_client(
+    server: FastMCP,
+    http_client_factory: HTTPClientFactoryFixture,
+) -> AsyncIterator[httpx.AsyncClient]:
+    async for client in http_client_factory(server):
+        yield client
